requirements.txt
fastapi
uvicorn
google-cloud-firestore
python-jose[cryptography]
pydantic
pydantic-settings
google-cloud-secret-manager
httpx
firebase-admin